pyproject.toml
[project]
name = "sub0-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arkiv-sdk>=1.0.0a8",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.2",
    "google-genai>=1.50.1",
    "greenlet>=3.2.4",
    "loguru>=0.7.3",
    "openai>=2.8.0",
    "pre-commit>=4.4.0",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
